+++ src/vs/workbench/services/suggest/browser/simpleSuggestWidgetRenderer.ts
@@ -60,7 +60,7 @@ export class SimpleSuggestWidgetItemRenderer implements IListRenderer<SimpleComp
 
 	readonly templateId = 'suggestion';
 
-	constructor(private readonly _getFontInfo: () => ISimpleSuggestWidgetFontInfo, @IConfigurationService private readonly _configurationService: IConfigurationService) {
+	constructor(private readonly _getFontInfo: () => ISimpleSuggestWidgetFontInfo, private readonly _onDidFontConfigurationChange: Event<void> | undefined, @IConfigurationService private readonly _configurationService: IConfigurationService) {
 	}
 
 	dispose(): void {
@@ -114,12 +114,15 @@ export class SimpleSuggestWidgetItemRenderer implements IListRenderer<SimpleComp
 		};
 
 		configureFont();
-
-		this._disposables.add(this._configurationService.onDidChangeConfiguration(e => {
-			if (e.affectsConfiguration('editor.fontSize') || e.affectsConfiguration('editor.fontFamily') || e.affectsConfiguration('editor.lineHeight') || e.affectsConfiguration('editor.fontWeight')) {
-				configureFont();
-			}
-		}));
+		if (this._onDidFontConfigurationChange) {
+			this._disposables.add(this._onDidFontConfigurationChange(() => configureFont()));
+		} else {
+			this._disposables.add(this._configurationService.onDidChangeConfiguration(e => {
+				if (e.affectsConfiguration('editor.fontSize') || e.affectsConfiguration('editor.fontFamily') || e.affectsConfiguration('editor.lineHeight') || e.affectsConfiguration('editor.fontWeight')) {
+					configureFont();
+				}
+			}));
+		}
 
 		return { root, left, right, icon, colorspan, iconLabel, iconContainer, parametersLabel, qualifierLabel, detailsLabel, disposables };
 	}
